Address comments, Change AutoLogin to LoginRepository
