+++ app/src/main/java/com/appdev/eateryblueandroid/util/LocationHandler.kt
@@ -5,6 +5,7 @@ import android.content.Context
 import android.content.pm.PackageManager
 import android.location.Location
 import androidx.core.app.ActivityCompat
+import com.appdev.eateryblueandroid.ui.appContext
 import com.appdev.eateryblueandroid.util.Constants.userPreferencesStore
 import com.codelab.android.datastore.PermissionSettings
 import com.google.android.gms.location.FusedLocationProviderClient
