+++ app/src/main/java/com/appdev/eateryblueandroid/util/NotificationsAndPrivacy.kt
@@ -1,5 +1,6 @@
 package com.appdev.eateryblueandroid.util
 
+import com.appdev.eateryblueandroid.ui.appContext
 import com.appdev.eateryblueandroid.util.Constants.userPreferencesStore
 import com.codelab.android.datastore.NotificationSettings
 import kotlinx.coroutines.CoroutineScope
